requirements.txt
pytesseract==0.3.10
Pillow>=10.1,<10.5
opencv-python==4.8.0.74
numpy==1.26.0
#flask==2.1.0  # If you intend to use a Flask-based server later for the webhook or streaming.
#requests==2.31.0  # For handling external requests, like webhooks.
easyocr
torch
torchvision